docs: Improve commands organization and docs
