--- src/commands/mod.rs
@@ -1,5 +1,0 @@
-mod prepare;
-mod query;
-
-pub use prepare::create_graph_from_dir;
-pub use query::get_affected_nodes;
